+++ src/components/pages/home/GetStarted.tsx
