+++ src/components/pages/home/Hero.tsx
