+++ src/components/pages/home/HowItWorks.tsx
